Feature Request #2744452 Make storing ParsedDocuments in CompressedCollection optional. Controlled by the parameter storeDocs, with a default of true. Usage: <storeDocs>false</storeDocs> to disable storage of the data.
Using this option will reduce the size of the indri repository for a given
collection by at least 1/2.

Doing so will not interfere with trec
evaluation query runs with IndriRunQuery, but not having the data in the
compressed collection may cause problems for applications that use the
ParsedDocument data structure, including the java retrieval gui.


git-svn-id: svn://svn.code.sf.net/p/lemur/code/indri/trunk@2142 14020d50-3050-45fe-8e0c-5ed7005ca764