IndriRunQuery: add stopfile commandline option
Add new option that allows for stopwords to be read from a file:

    IndriRunQuery -stopfile=stopfile.txt query.qry

where stopfile.txt contains each stopword on a single line.

The '-stopfile' option can be used in conjuction with the
'-stopper.word' option. Stopwords from both options are combined.

The '-stopfile' option appends the stopwords that are read from file
into the 'stopper.word' parameter field as if they were supplied on the
commandline or via the parameter file. This is because `QueryExpander`
accesses the stopwords from 'stopper.word' parameter field when stopping
is applied to the expanded query. It also allows for stopwords from both
commandline options to be combined.

Limitations

* IndriRunQuery (QueryEnvironment) only applies stopping to the expanded
  query, but not to the original query. See
  `QueryEnvironment::_runQuery` for more details.

Note: currently the '-stopfile' option only applies to IndriRunQuery, so
it can't be used in IndriBuildIndex for example.